Guard against group key duplication and mutation
The group key will get added to the rights metadata for a variety of
objects. If the group key would change for some reason all of those
object would need to be updated and reindexed. This seems like a lot of
work for updating an internal identifier. For now, I will punt and say
that group keys are immutable.