basic AccountStatus model and spec
This class will be the basis of what gets ActionCable'd to a dynamic
Status page.

The dynamic status page is a strategy for giving the user a place to
land while asynchronous account creation jobs are running.  It is also
a more valuable dashboard page on its own than the current HTML.